reorganized into folders and added draggable feature for folders
