+++ package.json
@@ -9,6 +9,7 @@
     "@types/react-dom": "^17.0.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-draggable": "^4.4.3",
     "react-scripts": "4.0.3",
     "typescript": "^4.2.4"
   },
