--- src/components/FooterComponent.tsx
@@ -1,75 +1,0 @@
-import React from "react";
-import '../styles/FooterComponent.css';
-export class FooterComponent extends React.Component {
-    
-    render() {
-        return (<React.Fragment>
-<div id="footerContainer">
-        <div className="imageContanier">
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-        </div>
-        <div className="imageContanier">
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img src='logo.png' alt='preeda logo' />
-            </div>
-            <div className="divImageContainer">
-            <img  src='logo.png' alt='preeda logo' />
-            </div>
-        </div>
-        </div>
-        </React.Fragment>)
-    }
-}
