--- src/components/MainBodyComponent.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import '../styles/MainBodyComponent.css';
-export const MainBodyComponent:React.FC = (): JSX.Element => {
-    return (
-        <React.Fragment>
-            <div id="mainBodyContainer">
-                <img src='address.png' alt="address"/>
-            </div>
-        </React.Fragment>
-    )
-}
