+++ src/components/footer/FooterComponent.tsx
@@ -1,0 +1,75 @@
+import React from "react";
+import './FooterComponent.css';
+export class FooterComponent extends React.Component {
+
+    render() {
+        return (<React.Fragment>
+            <div id="footerContainer">
+                <div className="imageContanier">
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                </div>
+                <div className="imageContanier">
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                </div>
+            </div>
+        </React.Fragment>)
+    }
+}
