+++ src/components/header/HeaderComponent.tsx
@@ -1,0 +1,75 @@
+import React from "react";
+import './HeaderComponent.css';
+export class HeaderComponent extends React.Component {
+
+    render() {
+        return (<React.Fragment>
+            <div id="container">
+                <div className="imageContanier">
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                </div>
+                <div className="imageContanier">
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                    <div className="divImageContainer">
+                        <img src='logo.png' alt='preeda logo' />
+                    </div>
+                </div>
+            </div>
+        </React.Fragment>)
+    }
+}
