--- src/styles/FolderComponent.css
@@ -1,29 +1,0 @@
-.desktop-item .title {
-    margin-top: 6px;
-    font-size: 12px;
-    text-align: center;
-    font-family: "SF Pro";
-    font-weight: 700;
-    color: #333;
-    text-shadow: 0 0 2px #fff, 0 0 7px #fff, 0 0 8px hsl(0deg 0% 100% / 30%);
-    transition: color .3s ease-in-out,text-shadow .3s ease-in-out;
-}
-
-.desktop-item .icon {
-    display: block;
-    margin: 0 auto;
-    width: 66px;
-    max-width: 100%;
-    transition: opacity 70ms ease-out;
-}
-
-
-.desktop-item {
-    width: 110px;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-    position: relative;
-    pointer-events: auto;
-}
