--- src/styles/MainBodyComponent.css
@@ -1,39 +1,0 @@
-#mainBodyContainer {
-    flex: 1;
-    overflow-x: hidden;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-@media (min-width: 300px) {
-    #mainBodyContainer {
-        flex:1;
-        overflow-x: hidden;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-    }
-    #mainBodyContainer img {
-        width: 28em;
-    }
-}
-
-@media (min-width: 1700px) {
-    #mainBodyContainer img {
-        width: 45em;
-    }
-}
-
-@media (min-width: 960px) {
-    #mainBodyContainer {
-        flex: 1;
-        overflow-x: hidden;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-    }
-    #mainBodyContainer img {
-        width: 45em;
-    }
-}
